[cleanup] adding documentation and moving some of the funtions
