+++ cmd/down.go
@@ -3,12 +3,12 @@ package cmd
 import (
 	"os"
 	"path/filepath"
-	"text/template"
 
 	"github.com/wawandco/transporter/Godeps/_workspace/src/github.com/codegangsta/cli"
 	"github.com/wawandco/transporter/utils"
 )
 
+//Down is the command runt when you do `tranporter down` on your CLI.
 func Down(ctx *cli.Context) {
 	temp := buildTempFolder()
 	defer os.RemoveAll(temp)
@@ -20,39 +20,14 @@ func Down(ctx *cli.Context) {
 		environment = ctx.Args()[0]
 	}
 
-	downTemplateData := CmdTemplateData{
+	downTemplateData := MainData{
 		TempDir:     temp,
 		Environment: environment,
 	}
 
 	commandArgs := utils.CopyMigrationFilesTo(temp)
-	main, _ := utils.WriteTemplateToFile(filepath.Join(temp, "main.go"), downTemplate, downTemplateData)
+	main, _ := utils.WriteTemplateToFile(filepath.Join(temp, "main.go"), utils.DownTemplate, downTemplateData)
 
 	commandArgs = append(commandArgs, main)
 	runTempFiles(commandArgs)
 }
-
-var downTemplate = template.Must(template.New("down.template").Parse(`
-package main
-
-import (
-	"log"
-	"path/filepath"
-	"github.com/wawandco/transporter/transporter"
-	"io/ioutil"
-)
-
-func main() {
-	log.Println("| Running Migrations Down on [{{.Environment}}] environment")
-	dat, _ := ioutil.ReadFile(filepath.Join("{{.TempDir}}","config.yml"))
-	db, err := transporter.DBConnection(dat, "{{.Environment}}")
-
-	if err != nil {
-		log.Println("Could not init database connection:", err)
-		return
-	}
-
-	defer db.Close()
-	transporter.RunOneMigrationDown(db)
-}
-`))
