+++ transporter/migration.go
@@ -13,10 +13,13 @@ type Migration struct {
 	Down       func(tx *sql.Tx)
 }
 
+//GetID Returns a string representation of the migration identifier.
 func (m *Migration) GetID() string {
 	return strconv.FormatInt(m.Identifier, 10)
 }
 
+//Pending returns if a particular migration is pending.
+// TODO: move it to be vendor specific inside the managers.
 func (m *Migration) Pending(db *sql.DB) bool {
 	rows, _ := db.Query("Select * from " + MigrationsTable + " WHERE identifier =" + m.GetID())
 	defer rows.Close()
@@ -29,6 +32,7 @@ func (m *Migration) Pending(db *sql.DB) bool {
 	return count == 0
 }
 
+//ByIdentifier is a specific order that causes first-created migrations to run first based on the identifier.
 type ByIdentifier []Migration
 
 func (a ByIdentifier) Len() int           { return len(a) }
