+++ utils/general.go
@@ -11,6 +11,7 @@ import (
 	"github.com/wawandco/transporter/Godeps/_workspace/src/github.com/crufter/copyrecur"
 )
 
+//WriteTemplateToFile writes a template result to a file.
 func WriteTemplateToFile(path string, t *template.Template, data interface{}) (string, error) {
 	f, e := os.Create(path)
 	if e != nil {
@@ -26,6 +27,8 @@ func WriteTemplateToFile(path string, t *template.Template, data interface{}) (s
 	return f.Name(), nil
 }
 
+//ReplaceInFile Replaces a text inside a file, this is usefull to replace the migrations package for the
+//main when copying the sources to run Up and Down.
 func ReplaceInFile(file, base, replacement string) {
 
 	input, err := ioutil.ReadFile(file)
@@ -49,6 +52,7 @@ func ReplaceInFile(file, base, replacement string) {
 	}
 }
 
+//CopyMigrationFilesTo copies the migration files into a specific folder, this is useful to run the Up and Down commands.
 func CopyMigrationFilesTo(tempFolder string) []string {
 	base := os.Getenv("TRANS_TESTING_FOLDER")
 	copyrecur.CopyFile(filepath.Join(base, "db", "config.yml"), filepath.Join(tempFolder, "config.yml"))
